+++ src/main/java/com/ecommerce/repository/ProductRepo.java
@@ -1,9 +1,8 @@
 package com.ecommerce.repository;
 
 import com.ecommerce.entity.Product;
-import com.ecommerce.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepo extends JpaRepository<Product,Integer> {
-    public Product findByPid(int pid);
+    public Product findByPid(Long pid);
 }
